refactor(api): use JpaRepository to implement findByAuthor in BookRepository

+++ backend/src/main/java/ir/ac/ut/ece/ie/books/BookPageDto.java
@@ -11,5 +11,5 @@ import java.util.List;
 @Data
 public class BookPageDto {
     List<BookDto> books;
-    int totalBooks;
+    Long totalBooks;
 }
